[REFACTOR] move spm related function to a separate package.
