+++ livingpark_utils/pipeline/PipelineABC.py
@@ -1,0 +1,5 @@
+from abc import ABC
+
+
+class PipelineABC(ABC):
+    pass
