+++ livingpark_utils/pipeline/exceptions.py
@@ -1,0 +1,7 @@
+"""Exceptions for pipeline executions."""
+
+
+class PipelineExecutionError(Exception):
+    """Exception class to report pipeline execution errors."""
+
+    pass
